Implemented first working test for permissions, fixes done:
- added missing `PermissionScope`: this must be used in permission maps
  when settings permissions
- using missing `.String()` on `PermissionAction` and `ResourceType` when
  building urls
- only the first func `PermissionOnResources` (the one tested) has been fixed
  for the time being

TODO: fix all the other permission methods + add missing tests

Refs #9

+++ custodia/api.go
@@ -11,7 +11,7 @@ import (
 // CustodiaEnvelope is the enveloped response, with data in subobject "data"
 type CustodiaEnvelope struct {
 	Result string `json:"result"`
-	ResultCode uint8 `json:"result_code"`
+	ResultCode uint64 `json:"result_code"`
 	Message json.RawMessage `json:"message"`
 	Data json.RawMessage `json:"data"`
 }
